requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
PyYAML>=6.0
pytz>=2021.1
numba>=0.53.0
scipy>=1.7.0
scikit-learn>=0.24.0
tqdm>=4.61.0
quantstats>=0.0.30  # For performance analytics
tabulate>=0.8.9  # For table formatting
pyarrow>=5.0.0   # For efficient data handling
plotly>=5.14.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0 